Redesigned the RegistryEngine (currently as RegistryEngine2) to simplyfy the usage and make it more intuitive.
Older RegistryEngine and related classes are still there and need to be
removed.

Now there can be many handler types, each concentrating on different
action categaries (resource related, tagging related, etc.)

InMemoryRegistry is renamed as a RegistryProvider, which provides registry storage
implementation to be used by the RegistryEngine.

RegistryEngine can use and provider implementing the RegistryProvider interface
as the storage.