Bunch of fixes to APP, code cleanup, etc.
Most of this *only* affects the RemoteRegistry/AtomRegistry, but the one noticeable other change is that doing a get(resourcePath + ";comments") to JDBCRegistry (i.e. getting a Resource of the comments) will now return a Resource with a Comment[] instead of a String[].  It was just silly to have to dereference another path just to get the comments, and we now have one rationalized way to return comments.  This way Atom comments will look right in a feed reader as well.

* Fix slug header support so we now POST to the parent collection and return the correct location header

* Placeholder sanitize method (we should decide what's allowed) for slugs

* Get rid of distinction between "$" and ";" - now there's just one separator.  Still need to remove all hardcoded references and use the constant.

* Introduce better error handling via StringResponseContext, which lets us send back stack traces on errors.

* Introduce an extension to note that a feed is a comments feed; this lets us return Comment[] remotely.

* Ratings feeds now contain the rating names and also the avg rating.  If avgRating is requested we only return that.

* Reimplement CommentCollectionURLHandler so it can just go ask the registry for the Comment[] and return that wrapped in a Resource.

* Get rid of System.out.println()s in tests!

* Usual code cleanup and spelling fixes