Implemented the resource content handling using streams. Now resource content can be set as an input stream, which is streamed to and from the database. In the registry webapp, the input stream taken from the request is written to the database without loading to the memory. This removes the limitation with the content size and now the limitation is only determined by the maximum blob size of the database. I tested this with 50 MB file and it worked fine with derby database.
byte[] based content handling is also supported for backword compatibility.
When resource.setContent(byteArray) is called, an input stream is created from the byte[] and set as the content input stream.
When resource.getContent() is called, input stream is loaded to memory and byte[] is returned.
Therefore, getContent may cause out of memory exceptions if the content is very large.
New applications should not call setContent and getContent methods and should call setContentStream and getContentStream instead.

APP registry is not completely modified to support stream based contents. Therefore, there are test failures in APP stuff.

Removed the hashmap based in memory registry as it is no longer used.