Some more jiggling of the Lifecycle stuff to get it closer to right.
Introduce SimpleLifecycleTest, which demonstrates that registering, associating, and using a very simple two-state lifecycle is working.

Use a multi-valued property (as Deepal was doing) to store associated Lifecycle names in the Resource.  I'm a little concerned about this approach since users can change property values - I think I'd prefer to actually bake lifecycles in to the data model, but this is OK for right now.

We need to think about whether we need an initialized RegistryContext even outside the realm of the RegistryServlet - what about the embedded case?

Considering we didn't have RegistryContext available, I added simple methods to JDBCRegistry to register and retrieve Lifecycles by name.

Also added a way to set a whole List for a multi-valued property.