updated the WSF/C build to get all the version info from a central location.
With this commit, the version information will be maintained in two files, init.bat and versions.in. Two files were necessary because we have used the version information in both batch files and nmake scripts. I think this is better than having them hard-coded everywhere, at least until we come up with a solution where we can do with only one file.