Get all APP tests running again (whew!) with Abdera 0.4.0.
Need to do a sweep through to clean everything up and revisit a few design points that came up during debugging, but this is substantially working now.

I noticed a problem with the APPBasedLifeCycleTest where permissions didn't seem to be getting copied when the resource was copied, resulting in a permission exception when trying to GET the resource after promotion - the problem happened in the regular LifeCycleTest when I switched the Registry to a SecureRegistry, so I commented both tests out for now.