Implemented the new resource store. This separates the versioning related operations from the normal resource operations.
Implemeneted transactions support. Now the transaction related methods provided in the Registry API can be used to perform 
multiple operations in a single transaction.

Added framework for fine grained permission model. Now all authorizations are performed in the components responsible for the actual operation.
This allows different components to impose different authorizations to match their exact requirement. This helps in the implementation
of pagination as well.

Introduced the user registry to track user's session. This is the replacement of SecureRegistry. But it only authenticates the user
and track user's name and authorization realm. Now all registry operations have to be performed through an instance of UserRegistry.

Introduced the embedded registry, which simplyfies the handling of user registries.

Changed the database schema to match the new resource store and the versioning model. All DAOs are updated to work with the new schema.
SQL related exception handling and logging is moved down to DAOs. Therefore, upper layers only have to deal with RegistryExceptions.

Moved resource content to a separate content table. Contents are stored against an UUID, so that duplication of contents can be
minimized.

Improved exception handling and logging.