More coalescing of initialization and configuration.
There will be one more step after this to finish cleaning up a bit more (which will involve renaming "EmbeddedRegistry" among other things), but this moves us a lot closer to a reasonable startup sequence.

The main thing users now do is create a RegistryContext.  You can hand the constructor an InputStream (with "registry.xml" info) or not (in which case it'll try to find registry.xml on the classpath).  You can also specify a datasource if you don't want to use one in registry.xml, and optionally a Realm.  Then you get your EmbeddedRegistry (i.e. factory) from the RegistryContext.