REFACTOR ALERT!!!
After conversation with Azeez and Deepal, switch order of parameters in addAssociation() and removeAssociation() to match rest of Registry methods (undoing the earlier change).  In particular, the first parameter is now the source path (i.e. the path being operated on which will receive or lose the association).

If you use these methods, please note that your code will CONTINUE to compile but will NOT work unless you change the order of your String params to match.

Also, get rid of outdated AtomRegistry class, and add mapping for /tags/*.