Whoops - mistakenly committed registry.xml last time.
Remove ReviewLifecycle (I'm still working on that), but check in Lifecycle.

This gives us a state-based lifecycle as discussed in today's meeting.  It's configured like this:

<aspect name="whatever" class="org.wso2.registry.aspects.Lifecycle">
 <state name="one"/>
 <state name="two>
  <condition property="color" condition="equals" value="red"/>
 </state>
 <state name="end"/>
</aspect>

This makes a lifecycle that has three states.  When you associate it with a resource it goes into the first state.  Promotion from the *second* state to the last one is conditional, based on the fact that a property named "color" has a particular value "red".  In addition to "equals" as a condition, we also support isNull, contains, greaterThan, and lessThan (the last two assume the property has an integer value).

Will be moving all this to modules/extensions in a bit.