Change ASSOCIATION table to support paths, not IDs.
This is for a couple of reasons - first, nothing was using the IDs at all.  All the outward Association APIs use paths, and no join queries were being made using the IDs.  Second, this allows us to support associations that go to/from resources outside our Registry URL space (a source/target path does not have to map to our resource ID).  So now you can have an association which is of type "commentary" and points to someone's external blog entry page.

I already changed the UI to support this a few days ago - so notice you can now see this working with this change.

This is a critical feature for us into the future, so making the db change (which affects no external-facing APIs) now.