MASHUP-972 fix.  Seems to have to do with custom UIs named index.htm.
It looked like the "this" context of the nested check for "index.htm" is incorrectly set (though Firebug acts pretty weird on FF3 so maybe that's a debugging artifact.)  In any case I check for index.html and index.htm in parallel now instead of in sequence, and I haven't seen a recurrence.  Tyrell, please review the code since this appears to have lots of tricky failover behavior.