Applying the following fix to the branch
Author: keith Date: Mon Sep 1 23:27:05 2008 New Revision: 21318 URL: http://wso2.org/svn/browse/wso2?view=rev&revision=21318

Log: Fixing Mashup-1101. Introduced a ProxyConfiguration class that extends the one
in axis2 in order to configure the proxy details. The one in axis2 relied on the
message Context to configure it which is not available in the Mashup Server on some instances. Hence put in a new
method so that we can configure it using the configuration context.

Updating the places where ProxyCopnfiguration was used to use this newly added
class.