Modified the CarbonAxisConfigurator to create transport descriptions for the transports persisted in the registry. This makes sure that when the AxisConfiguration is first created, all the transport descriptions are properly set. As a result all the service deployments should succeed without any issues.
With this change in place we no longer need to create transport descriptions in the CarbonServerManager. (This is not the right place to do it and most of the time it is too late anyway to init transports at that level.) For the moment I have simply commented out the necessary method call in the CarbonServerManager. After some testing we can get rid of that code.